+++ src/Core/Language/Utf8/IDocumentHashProvider.cs
@@ -4,6 +4,8 @@ namespace HotChocolate.Language
 {
     public interface IDocumentHashProvider
     {
+        string Name { get; }
+
         string ComputeHash(ReadOnlySpan<byte> document);
     }
 }