+++ src/DataLoader/Benchmarks/CompoundKey.cs
@@ -1,0 +1,8 @@
+namespace GreenDonut.Benchmarks
+{
+    public class CompoundKey
+    {
+        public int Integer { get; set; }
+        public string String { get; set; }
+    }
+}